requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
Flask==2.1.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.0
psycopg2-binary==2.9.9
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==1.3.24

